Finish major refactor
